Triggers and automation for expiring coin and locked coin with Wallet time-based integrateion test (#2279)
* Triggers and automation for expiring coin and locked coin

* format

* address comments

* add Coin and LockedCoin in svc acs store contract filter

* format

* testing coin and locked expiry

* testing coin and locked expiry

* split the test cases for testing coin and locked coin expiry

* Update apps/app/src/test/scala/com/daml/network/integration/tests/WalletTimeBasedIntegrationTest.scala


* Update apps/svc/src/main/scala/com/daml/network/svc/store/SvcStore.scala


* address comments

* Fire debug log instead of warning for CoinArchive of CoinContract due to SVC automation

* merge

* fix xml

+++ apps/app/src/test/scala/com/daml/network/util/WalletTestUtil.scala
@@ -80,7 +80,7 @@ trait WalletTestUtil extends CoinTestCommon with CnsTestUtil {
       userParty: PartyId,
       validatorParty: PartyId,
       coins: Seq[GrpcWalletAppClient.CoinPosition],
-      quantity: Int,
+      quantity: BigDecimal,
       transferContext: v1.coin.AppTransferContext,
       expiredDuration: Duration,
   )(implicit coinEnv: CoinTestConsoleEnvironment): Unit =
@@ -113,7 +113,7 @@ trait WalletTestUtil extends CoinTestCommon with CnsTestUtil {
                   ).asJava,
                   Seq[v1.coin.TransferOutput](
                     new v1.coin.transferoutput.OutputSenderCoin(
-                      Some(BigDecimal(quantity).bigDecimal).toJava,
+                      Some(quantity.bigDecimal).toJava,
                       Some(
                         new v1.coin.TimeLock(
                           Seq(userParty.toProtoPrimitive).asJava,
